[Feat/#12] 유저 인증 메서드 클래스 생성
