+++ ABloom/ABloom/Login/LoginView.swift
@@ -33,6 +33,10 @@ struct LoginView: View {
       .padding(.horizontal, 20)
       .padding(.bottom, 60)
     }
+    .task {
+      try? loginVM.loadCurrentUser()
+      print(loginVM.user)
+    }
   }
 }
